requirements.txt
streamlit
torch
torchvision
Pillow
transformers
peft
datasets
evaluate
numpy
